factor out mock chat values
