+++ src/redux/root-reducer.test.ts
@@ -1,6 +1,6 @@
 import { rootReducer, Actions } from './root-reducer';
 import { GlobalState } from '../shared-types/global-state';
-import { dummyCurrentUser } from './mock-messages';
+import { dummyCurrentUser } from '../common/mock-messages';
 import { User } from '../clients/user-data';
 
 const activeChatId = '123';
