+++ src/shared-types/global-state.ts
@@ -1,6 +1,6 @@
 import { Chats } from '../clients/messages-data';
 import { User } from '../clients/user-data';
-import { mockChats, dummyCurrentUser, chatId1 } from '../redux/mock-messages';
+import { mockChats, dummyCurrentUser, chatId1 } from '../common/mock-messages';
 
 export type GlobalState = {
   activeChatId: string | null;
